js: favor Crankshaft over TurboFan
Crankshaft is currently better at optimizing our 'compute' function than
its successor TurboFan.

It's probably just a matter of time before TurboFan takes over but in
the meantime, let's make our 'compute' function Crankshaft-optimizable
by getting rid of the 'for...of' statement and moving let assignments at
the beginning of the scope.

This results in 4x speed-up under Node v5.0.0.

On a 2015 MacBook:

    $ ruby bench.rb csa.js
    Average time per search: 23 ms